out_stackdriver: implement the second feature gap - insertId
